refactor: inbox 코드 구조 정리
