+++ apps/web/src/components/GNB/Notification/InboxList.tsx
@@ -1,66 +1,17 @@
-'use client';
-
 /* eslint-disable @next/next/no-img-element */
-import { useState } from 'react';
 import { useTranslations } from 'next-intl';
 import { css, cx } from '_panda/css';
 import { center } from '_panda/patterns';
 import type { Inbox } from '@gitanimals/api';
 import { inboxQueries, useReadInbox } from '@gitanimals/react-query';
-import { wrap } from '@suspensive/react';
-import { useQueryClient, useSuspenseQuery } from '@tanstack/react-query';
-import { BellIcon } from 'lucide-react';
+import { useQueryClient } from '@tanstack/react-query';
 
 import { Link } from '@/i18n/routing';
 import { customScrollStyle } from '@/styles/scrollStyle';
 
 import { AnimatePortal } from '../../Portal';
 
-import { INBOX_NOTICE } from './inbox.contants';
-
-export const Notification = wrap
-  .ErrorBoundary({ fallback: <BellIcon size={24} /> })
-  .Suspense({ fallback: <BellIcon size={24} /> })
-  .on(function Notification() {
-    const t = useTranslations('Inbox');
-    const [isOpen, setIsOpen] = useState(false);
-    const { data } = useSuspenseQuery(inboxQueries.getAllUnreadInboxOptions());
-
-    const inboxCount = data.inboxes.filter((inbox) => inbox.status === 'UNREAD').length;
-    const inboxList = [
-      ...INBOX_NOTICE.map((item) => ({ ...item, title: t(item.title), body: t(item.body) })),
-      ...data.inboxes,
-    ];
-
-    return (
-      <div>
-        <button className={notificationStyle} onClick={() => setIsOpen((prev) => !prev)}>
-          <BellIcon size={24} />
-          {inboxCount > 0 && <div className={countStyle}>{inboxCount}</div>}
-        </button>
-
-        <InboxList isOpen={isOpen} list={inboxList} />
-      </div>
-    );
-  });
-
-const notificationStyle = css({
-  position: 'relative',
-});
-
-const countStyle = center({
-  borderRadius: '12px',
-  background: '#019C5A',
-  position: 'absolute',
-  top: '-10px',
-  right: '-10px',
-  textStyle: 'glyph12.regular',
-  width: '18px',
-  height: '18px',
-  color: 'white',
-});
-
-const InboxList = ({ isOpen, list }: { isOpen: boolean; list: Inbox[] }) => {
+export const InboxList = ({ isOpen, list }: { isOpen: boolean; list: Inbox[] }) => {
   const t = useTranslations('Layout');
 
   return (
